move create lobby modal to component
