+++ src/pages/index.tsx
@@ -3,15 +3,11 @@ import Button from "@ui/Button";
 import Input from "@ui/Input";
 import Label from "@ui/Label";
 import { LoadingSpinner } from "@ui/Loader";
-import Modal from "@ui/Modal";
-import { useForm, SubmitHandler } from "react-hook-form";
 import { type NextPage } from "next";
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
-import { trpc } from "@utils/trpc";
-import { socket } from "@utils/websocket/socket";
-import { useGameStore } from "@utils/zustand/gameStore";
+import { useState } from "react";
+import CreateLobbyModal from "@components/playQuiz/CreateLobbyModal";
 
 interface IFormInputs {
   quiz: string;
@@ -23,51 +19,6 @@ const Home: NextPage = () => {
   const { data: session, status } = useSession();
   const router = useRouter();
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const setSocketId = useGameStore((state) => state.setSocketId);
-  const setUserName = useGameStore((state) => state.setUserName);
-  const setQuizName = useGameStore((state) => state.setQuizName);
-  const setNumOfQuestions = useGameStore((state) => state.setNumOfQuestions);
-  const setNumOfTeams = useGameStore((state) => state.setNumOfTeams);
-
-  const getQuizzes = trpc.quiz.getAllQuizzes.useQuery();
-
-  const createLobby = (data: IFormInputs) => {
-    const quiz = JSON.parse(data.quiz);
-
-    setIsLoading(true);
-    setUserName(data.userName);
-    setSocketId(socket.id);
-    setQuizName(quiz.title);
-    setNumOfQuestions(quiz.numOfQuestions);
-    setNumOfTeams(data.numOfTeams);
-    socket.emit(
-      "create-lobby",
-      {
-        quizId: quiz.id,
-        userName: data.userName,
-        numOfTeams: data.numOfTeams,
-        quizName: quiz.title,
-        numOfQuestions: quiz.numOfQuestions,
-      },
-      (response: { id: string }) => {
-        setIsLoading(false);
-        router.push(`/play/${response.id}`);
-      }
-    );
-  };
-
-  useEffect(() => {
-    getQuizzes.refetch();
-  }, []);
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<IFormInputs>();
-  const onSubmit: SubmitHandler<IFormInputs> = (data) => createLobby(data);
 
   return (
     <>
@@ -110,6 +61,10 @@ const Home: NextPage = () => {
                       Create Room
                     </button>
                   </div>
+                  <CreateLobbyModal
+                    isModalOpen={isModalOpen}
+                    setIsModalOpen={setIsModalOpen}
+                  />
                 </>
               ) : (
                 <>
@@ -128,62 +83,6 @@ const Home: NextPage = () => {
             </div>
           </div>
         )}
-        <Modal open={isModalOpen} setOpen={setIsModalOpen}>
-          <form onSubmit={handleSubmit(onSubmit)}>
-            <div className="flex w-80 flex-col gap-6">
-              <div className="flex flex-col gap-2">
-                <Label text="Username" />
-                <input
-                  placeholder="Enter Username"
-                  maxLength={30}
-                  {...register("userName", { required: true })}
-                  className="w-full rounded-md border border-transparent bg-zinc-700 p-3 text-base font-medium text-zinc-200 focus:outline-none"
-                  autoFocus
-                />
-                {errors.userName && (
-                  <div className="text-red-400">Username is required</div>
-                )}
-              </div>
-              <div className="flex flex-col items-center gap-2">
-                <Label text="Quiz" />
-                {getQuizzes.isLoading ? (
-                  <LoadingSpinner />
-                ) : (
-                  <select
-                    className="w-full rounded-md border border-transparent bg-zinc-700 p-3 text-base text-zinc-200 focus:outline-none"
-                    {...register("quiz", { required: true })}
-                  >
-                    {getQuizzes.data?.map((quiz) => (
-                      <option
-                        key={quiz.id}
-                        value={`{"id": "${quiz.id}","title": "${
-                          quiz.title
-                        }","numOfQuestions": "${quiz._count.questions.toString()}"}`}
-                      >
-                        {quiz.title}
-                      </option>
-                    ))}
-                  </select>
-                )}
-              </div>
-              <div className="flex flex-col gap-2">
-                <Label text="Number of teams" />
-                <select
-                  className="w-full rounded-md border border-transparent bg-zinc-700 p-3 text-base text-zinc-200 focus:outline-none"
-                  {...register("numOfTeams", { required: true })}
-                >
-                  <option>2</option>
-                  <option>3</option>
-                  <option>4</option>
-                  <option>5</option>
-                </select>
-              </div>
-              <Button type="submit" disabled={isLoading}>
-                {isLoading ? <LoadingSpinner /> : "Create"}
-              </Button>
-            </div>
-          </form>
-        </Modal>
       </main>
     </>
   );
